Move version lookup to a shared helper
