+++ spec/services/tasks/pubmed_ingest/recurring/utilities/file_attachment_service_spec.rb
@@ -270,58 +270,6 @@ RSpec.describe Tasks::PubmedIngest::Recurring::Utilities::FileAttachmentService 
     end
   end
 
-  describe '#latest_version_prefix' do
-    let(:list_xml) do
-      <<~XML
-        <?xml version="1.0" encoding="UTF-8"?>
-        <ListBucketResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
-          <CommonPrefixes><Prefix>PMC123456.1/</Prefix></CommonPrefixes>
-          <CommonPrefixes><Prefix>PMC123456.2/</Prefix></CommonPrefixes>
-        </ListBucketResult>
-      XML
-    end
-
-    before do
-      allow(HTTParty).to receive(:get)
-        .with(a_string_including('PMC123456.'), anything)
-        .and_return(double('response', code: 200, body: list_xml))
-    end
-
-    it 'returns the latest (sorted) version prefix' do
-      expect(service.latest_version_prefix('PMC123456')).to eq('PMC123456.2/')
-    end
-
-    context 'when no versions exist' do
-      let(:empty_xml) do
-        <<~XML
-          <?xml version="1.0" encoding="UTF-8"?>
-          <ListBucketResult xmlns="http://s3.amazonaws.com/doc/2006-03-01/">
-          </ListBucketResult>
-        XML
-      end
-
-      before do
-        allow(HTTParty).to receive(:get)
-          .and_return(double('response', code: 200, body: empty_xml))
-      end
-
-      it 'returns nil' do
-        expect(service.latest_version_prefix('PMC123456')).to be_nil
-      end
-    end
-
-    context 'when the S3 listing request fails' do
-      before do
-        allow(HTTParty).to receive(:get)
-          .and_return(double('response', code: 500, body: ''))
-      end
-
-      it 'raises an error' do
-        expect { service.latest_version_prefix('PMC123456') }.to raise_error(/S3 listing failed/)
-      end
-    end
-  end
-
   describe '#fetch_s3_file' do
     let(:url) { 'https://pmc-oa-opendata.s3.amazonaws.com/PMC123456.1/PMC123456.1.pdf' }
     let(:local_path) { '/tmp/test_fulltext/PMC123456_001.pdf' }
@@ -344,6 +292,12 @@ RSpec.describe Tasks::PubmedIngest::Recurring::Utilities::FileAttachmentService 
 
       expect(result).to eq(404)
       expect(File).not_to have_received(:binwrite)
+    end
+  end
+
+  describe 'module composition' do
+    it 'includes the shared PMC S3 version lookup helper' do
+      expect(described_class.included_modules).to include(Tasks::PubmedIngest::SharedUtilities::PmcS3VersionLookup)
     end
   end
 end
